requirements.txt
https://github.com/thesadru/genshin.py/archive/refs/heads/dev.zip
python-dotenv
asyncio
aiohttp
rsa